+++ mod.ts
@@ -82,3 +82,4 @@ export { HEICFormat } from "./src/formats/heic.ts";
 export { AVIFFormat } from "./src/formats/avif.ts";
 
 export { decodeBase64, encodeBase64, parseDataUrl, toDataUrl } from "./src/utils/base64.ts";
+export { cmykToRgb, cmykToRgba, rgbaToCmyk, rgbToCmyk } from "./src/utils/image_processing.ts";